made backlight a wiget, edited based on other comments
