+++ src/i19serial_ui/gui/widgets/__init__.py
@@ -1,7 +1,15 @@
+from .backlight_box import BacklightBox
 from .grid_options import GridOptions
 from .input_panel import InputPanel
 from .log_box import LogBox
 from .phi_rotator import PhiAdjust
 from .wells_selection import WellsSelectionPanel
 
-__all__ = ["InputPanel", "GridOptions", "WellsSelectionPanel", "LogBox", "PhiAdjust"]
+__all__ = [
+    "InputPanel",
+    "GridOptions",
+    "WellsSelectionPanel",
+    "LogBox",
+    "PhiAdjust",
+    "BacklightBox",
+]
